--- a/build.sbt
+++ b/build.sbt
@@ -8,12 +8,12 @@ lazy val buildSettings = Seq(
   crossScalaVersions := Seq("2.11.12", "2.12.6")
 )
 
+lazy val arrowsVersion = "0.1.20-SNAPSHOT"
 lazy val finagleVersion = "18.7.0"
 lazy val twitterServerVersion = "18.7.0"
 lazy val circeVersion = "0.9.3"
 lazy val circeIterateeVersion = "0.9.0"
 lazy val catbirdVersion = "18.6.0"
-lazy val arrowsVersion = "0.1.20-SNAPSHOT"
 lazy val shapelessVersion = "2.3.3"
 lazy val catsVersion = "1.1.0"
 lazy val sprayVersion = "1.3.4"