--- a/build.sbt
+++ b/build.sbt
@@ -8,20 +8,20 @@ lazy val buildSettings = Seq(
   crossScalaVersions := Seq("2.11.12", "2.12.6")
 )
 
-lazy val arrowsVersion = "0.1.20-SNAPSHOT"
-lazy val finagleVersion = "18.7.0"
-lazy val twitterServerVersion = "18.7.0"
-lazy val circeVersion = "0.9.3"
-lazy val circeIterateeVersion = "0.9.0"
-lazy val catbirdVersion = "18.6.0"
-lazy val shapelessVersion = "2.3.3"
-lazy val catsVersion = "1.1.0"
-lazy val sprayVersion = "1.3.4"
 lazy val argonautVersion = "6.2.2"
-lazy val json4sVersion = "3.5.4"
-lazy val iterateeVersion = "0.17.0"
+lazy val arrowsVersion = "0.1.20-SNAPSHOT"
+lazy val catbirdVersion = "18.6.0"
+lazy val catsVersion = "1.1.0"
+lazy val circeIterateeVersion = "0.9.0"
+lazy val circeVersion = "0.9.3"
+lazy val finagleVersion = "18.7.0"
 lazy val iterateeTwitterVersion = "18.7.0"
+lazy val iterateeVersion = "0.17.0"
+lazy val json4sVersion = "3.5.4"
 lazy val refinedVersion = "0.9.2"
+lazy val shapelessVersion = "2.3.3"
+lazy val sprayVersion = "1.3.4"
+lazy val twitterServerVersion = "18.7.0"
 
 lazy val compilerOptions = Seq(
   "-deprecation",