--- a/build.sbt
+++ b/build.sbt
@@ -8,7 +8,7 @@ lazy val buildSettings = Seq(
 )
 
 lazy val argonautVersion = "6.2.2"
-lazy val arrowsVersion = "0.1.20-SNAPSHOT"
+lazy val arrowsVersion = "0.1.21"
 lazy val catbirdVersion = "18.7.0"
 lazy val catsVersion = "1.1.0"
 lazy val circeIterateeVersion = "0.9.0"