--- a/core/src/main/scala/io/finch/endpoint/body.scala
+++ b/core/src/main/scala/io/finch/endpoint/body.scala
@@ -3,11 +3,13 @@ package io.finch.endpoint
 import com.twitter.concurrent.AsyncStream
 import com.twitter.io.Buf
 import com.twitter.util.{Future, Return, Throw, Try}
-import io.catbird.util.Rerunnable
 import io.finch._
 import io.finch.internal._
 import io.finch.items._
 import java.nio.charset.Charset
+
+import arrows.twitter.Task
+
 import scala.reflect.ClassTag
 
 private abstract class FullBody[A] extends Endpoint[A] {
@@ -18,7 +20,7 @@ private abstract class FullBody[A] extends Endpoint[A] {
   final def apply(input: Input): Endpoint.Result[A] =
     if (input.request.isChunked) EndpointResult.NotMatched
     else {
-      val output = Rerunnable.fromFuture {
+      val output = Task.async {
         val contentLength = input.request.contentLengthOrNull
         if (contentLength == null || contentLength == "0") missing
         else present(input.request.content, input.request.charsetOrUtf8)
@@ -168,7 +170,7 @@ private[finch] trait Bodies {
         EndpointResult.Matched(
           input,
           Trace.empty,
-          Rerunnable(Output.payload(AsyncStream.fromReader(input.request.reader)))
+          Task(Output.payload(AsyncStream.fromReader(input.request.reader)))
         )
 
     final override def item: RequestItem = items.BodyItem