--- a/core/src/main/scala/io/finch/endpoint/header.scala
+++ b/core/src/main/scala/io/finch/endpoint/header.scala
@@ -1,10 +1,11 @@
 package io.finch.endpoint
 
+import arrows.twitter.Task
 import cats.Id
 import com.twitter.util.{Future, Return, Throw, Try}
-import io.catbird.util.Rerunnable
 import io.finch._
 import io.finch.items._
+
 import scala.reflect.ClassTag
 
 private abstract class Header[F[_], A](
@@ -22,7 +23,7 @@ private abstract class Header[F[_], A](
   }
 
   final def apply(input: Input): Endpoint.Result[F[A]] = {
-    val output = Rerunnable.fromFuture {
+    val output = Task.async {
       input.request.headerMap.getOrNull(name) match {
         case null => missing(name)
         case value => Future.const(d(value).transform(self))