… core/src/main/scala/io/finch/endpoint/multipart.scala
… core/src/main/scala/io/finch/internal/package.scala
… core/src/test/scala/io/finch/FinchSpec.scala
… iteratee/src/main/scala/io/finch/iteratee/package.scala
… project/plugins.sbt