A first attempt to solve #954, learning Arrow/Task on the side. Code compiles but tests fail with `java.lang.NoSuchMethodError: com.twitter.util.Local$.save()[Lscala/Option;`

Comments welcome