@vkostyukov 2018-07-13
Really cool!  @erikwj 

I bet you see a failure b/c we recently changed the API for Future Locals in Finagle but Arrows project hasn't updated yet to the most recent version of util.