@erikwj 2018-07-14
Thanks for your comments @vkostyukov I'll continue with it next week