@erikwj 2018-07-18
Updating to twitter-util 18.7.0 helped a lot