@vkostyukov 2018-07-18
Hey @erikwj! This looks pretty decent already! What's failing?

I looked at the CI job and it seems like scalastyle isn't super happy. You can validate the style (and everything else) locally with `sbt validate`.