@vkostyukov 2018-07-18
No, thank you @erikwj! I think the most exciting part is to run some benchmarks and see how `Task` compares to `Returnable`. Do you think you can try it?

I'd suggest to run out wrk benchmark as a first step to make sure the big picture still looks good. After that we can narrow it down to more specific workloads: there is a couple of `Bootstrap*` benchamrks we can run as well.