@erikwj 2018-07-18
Yes, I'll give it a try, but probably on Friday