@fwbrasil 2018-07-18
LGTM. The new release is available: http://search.maven.org/#artifactdetails%7Cio.trane%7Carrows-twitter_2.11%7C0.1.21%7Cjar