@travisbrown 2018-07-19
Very simple benchmark results here (in a WIP PR for the catbird instances for arrows): https://github.com/travisbrown/catbird/pull/70