@vkostyukov 2018-07-19
Here is a sneak peek into the product (`::`) benchmark results (the difference is pretty bizarre):

```
BEFORE:

[info] Benchmark                                                   Mode  Cnt     Score     Error   Units
[info] ProductBenchmark.bothMatched                                avgt   10   984.884 ±  35.849   ns/op
[info] ProductBenchmark.bothMatched:·gc.alloc.rate.norm            avgt   10  1472.000 ±   0.001    B/op
[info] ProductBenchmark.leftMatched                                avgt   10    27.908 ±   1.950   ns/op
[info] ProductBenchmark.leftMatched:·gc.alloc.rate.norm            avgt   10    80.000 ±   0.001    B/op
[info] ProductBenchmark.rightMatched                               avgt   10     5.059 ±   0.179   ns/op
[info] ProductBenchmark.rightMatched:·gc.alloc.rate.norm           avgt   10    ≈ 10⁻⁶              B/op


AFTER

[info] ProductBenchmark.bothMatched                                avgt   10   361.867 ±   8.955   ns/op
[info] ProductBenchmark.bothMatched:·gc.alloc.rate.norm            avgt   10   712.000 ±  12.749    B/op
[info] ProductBenchmark.leftMatched                                avgt   10    29.446 ±   2.453   ns/op
[info] ProductBenchmark.leftMatched:·gc.alloc.rate.norm            avgt   10    80.000 ±   0.001    B/op
[info] ProductBenchmark.rightMatched                               avgt   10     5.152 ±   0.326   ns/op
[info] ProductBenchmark.rightMatched:·gc.alloc.rate.norm           avgt   10    ≈ 10⁻⁶              B/op
```