@vkostyukov 2018-07-19
Here is the coproduct (`:+:`) benchmark:

```
BEFORE:

[info] Benchmark                                                      Mode  Cnt     Score     Error   Units
[info] CoproductBenchmark.bothMatched                                 avgt   10   242.465 ±  11.592   ns/op
[info] CoproductBenchmark.bothMatched:·gc.alloc.rate.norm             avgt   10   424.000 ±   0.001    B/op
[info] CoproductBenchmark.leftMatched                                 avgt   10   209.721 ±  15.328   ns/op
[info] CoproductBenchmark.leftMatched:·gc.alloc.rate.norm             avgt   10   344.000 ±   0.001    B/op
[info] CoproductBenchmark.rightMatched                                avgt   10   204.683 ±   6.273   ns/op
[info] CoproductBenchmark.rightMatched:·gc.alloc.rate.norm            avgt   10   336.000 ±  12.749    B/op

AFTER:

[info] CoproductBenchmark.bothMatched                                 avgt   10    83.733 ±   1.724   ns/op
[info] CoproductBenchmark.bothMatched:·gc.alloc.rate.norm             avgt   10   272.000 ±   0.001    B/op
[info] CoproductBenchmark.leftMatched                                 avgt   10    49.858 ±   1.122   ns/op
[info] CoproductBenchmark.leftMatched:·gc.alloc.rate.norm             avgt   10   168.000 ±   0.001    B/op
[info] CoproductBenchmark.rightMatched                                avgt   10    50.779 ±   1.250   ns/op
[info] CoproductBenchmark.rightMatched:·gc.alloc.rate.norm            avgt   10   168.000 ±   0.001    B/op
```