@vkostyukov 2018-07-23
I changed the base branch from `master` to `arrows` as we want to release this a part of its own artifact: `finch-arrows-x`.